sockets/backend/soul/requirements.txt
djangorestframework==3.15.2
pillow==11.1.0
pytubefix==8.12.2
drf-spectacular==0.28.0
psycopg-binary==3.2.6
django-cors-headers==4.7.0